refactor: 테스트와 support 분리
